# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Continua API
  version: 0.1.0
  description: AI Agent Observability Platform API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
    description: Relative path (works with any host)
security:
  - auth0Bearer: []
  - apiKey: []
paths:
...
  /api/traces:
    get:
      operationId: listTraces
      summary: List traces
      tags: [Traces]
      parameters:
        - $ref: '#/components/parameters/SelectedProjectId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: session_id
          in: query
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          description: Full-text search query (searches trace name and user_id)
          schema:
            type: string
        - name: sort_by
          in: query
          description: Sort traces by started time. Ignored when q is present.
          schema:
            type: string
            enum: [started_at]
        - name: sort_dir
          in: query
          description: Sort direction for started time. Ignored when q is present.
          schema:
            type: string
            enum: [asc, desc]
        - name: status
          in: query
          description: Filter by trace status
          schema:
            type: string
            enum: [running, completed, failed]
        - name: start_time_from
          in: query
          description: Filter traces starting at or after this time
          schema:
            type: string
            format: date-time
        - name: start_time_to
          in: query
          description: Filter traces starting at or before this time
          schema:
            type: string
            format: date-time
        - name: user_id
          in: query
          description: Filter by user ID
          schema:
            type: string
        - name: engine_instance_key
          in: query
          description: Filter by engine instance key
          schema:
            type: string
        - name: engine_run_id
          in: query
          description: Filter by engine run ID
          schema:
            type: string
            format: uuid
        - name: engine_definition_name
          in: query
          description: Filter by engine definition name
          schema:
            type: string
        - name: engine_definition_version
          in: query
          description: Filter by engine definition version
          schema:
            type: string
        - name: engine_run_status
          in: query
          description: Filter by engine run lifecycle status
          schema:
            type: string
            enum: [queued, running, waiting, suspended, completed, failed, cancelled, terminated, continued_as_new]
        - name: engine_parent_run_id
          in: query
          description: Filter by engine parent run ID
          schema:
            type: string
            format: uuid
        - name: engine_root_run_id
          in: query
          description: Filter by engine root run ID
          schema:
            type: string
            format: uuid
        - name: engine_child_key
          in: query
          description: Filter by engine child key
          schema:
            type: string
        - name: engine_child_depth
          in: query
          description: Filter by engine child depth
          schema:
            type: integer
        - name: engine_projection_state
          in: query
          description: Filter by engine projection state
          schema:
            $ref: '#/components/schemas/EngineProjectionState'
        - name: has_errors
          in: query
          description: Filter traces with errors (error_count > 0)
          schema:
            type: boolean
        - name: min_duration_ms
          in: query
          description: Filter traces with duration >= this value in milliseconds
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of traces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceList'
        '401':
          description: Unauthorized - missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/traces/{id}:
    get:
      operationId: getTrace
      summary: Get a trace by ID
      tags: [Traces]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SelectedProjectId'
      responses:
        '200':
          description: Trace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceDetail'
        '401':
          description: Unauthorized - missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/traces/{id}/spans:
    get:
      operationId: listSpansByTrace
      summary: List spans for a trace
      tags: [Spans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SelectedProjectId'
      responses:
        '200':
          description: List of spans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpanList'
        '401':
          description: Unauthorized - missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/traces/{id}/events:
    get:
      operationId: getTraceEvents
      summary: List timeline events for a trace
      tags: [Traces]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SelectedProjectId'
        - name: after
          in: query
          description: Opaque cursor returned by a previous timeline response
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of timeline events to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Timeline events for the trace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineResponse'
        '400':
          description: Invalid cursor or pagination parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    SelectedProjectId:
      name: project_id
      in: query
      description: Required when authenticating with an Auth0 bearer token. Ignored when the request is already project-scoped by API key.
      schema:
        type: string
        format: uuid
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    auth0Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 access token for debugger operator authentication
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    EngineProjectionState:
      type: string
      enum: [up_to_date, catching_up, summary_only, journal_expired]
...
    EngineRunStatus:
      type: string
      enum: [QUEUED, RUNNING, WAITING, SUSPENDED, COMPLETED, FAILED, CANCELLED, TERMINATED, CONTINUED_AS_NEW]
    EngineTraceInfo:
      type: object
      required: [run_id, definition_name, definition_version, projection_state]
      properties:
        run_id:
          type: string
          format: uuid
        definition_name:
          type: string
        definition_version:
          type: string
        projection_state:
          $ref: '#/components/schemas/EngineProjectionState'
        parent_run_id:
          type: string
          format: uuid
        root_run_id:
          type: string
          format: uuid
        child_key:
          type: string
        child_depth:
          type: integer
    EngineWaitState:
      type: object
      additionalProperties: true
      properties:
        kind:
          type: string
        activity_key:
          type: string
        activity_type:
          type: string
        timer_key:
          type: string
        due_at:
          type: string
          format: date-time
        signal_name:
          type: string
        child_key:
          type: string
    EnginePendingWork:
      type: object
      required: [pending_activity_tasks, pending_inbox_items]
      properties:
        pending_activity_tasks:
          type: integer
          format: int64
        pending_inbox_items:
          type: integer
          format: int64
...
    EngineFailureSummary:
      type: object
      required: [error_code, error_message, status]
      properties:
        error_code:
          type: string
          description: |
            Stable reserved value: `definition_version_mismatch`. Clients may
            exact-match this value while still accepting arbitrary strings.
        error_message:
          type: string
        status:
          type: string
    EngineRunSummary:
      allOf:
        - $ref: '#/components/schemas/EngineTraceInfo'
        - type: object
          required: [status, instance_key, pending_work, created_at, updated_at]
          properties:
            status:
              $ref: '#/components/schemas/EngineRunStatus'
            instance_key:
              type: string
            continued_from_run_id:
              type: string
              format: uuid
            continued_to_run_id:
              type: string
              format: uuid
            continued_from_trace_id:
              type: string
            continued_to_trace_id:
              type: string
            custom_status:
              type: object
            wait_state:
              $ref: '#/components/schemas/EngineWaitState'
            pending_work:
              $ref: '#/components/schemas/EnginePendingWork'
            result:
              description: Terminal workflow result payload when available.
            failure:
              $ref: '#/components/schemas/EngineFailureSummary'
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            completed_at:
              type: string
              format: date-time
    Trace:
      type: object
      required: [id, name, status, started_at]
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        session_external_id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [RUNNING, COMPLETED, FAILED]
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        total_tokens_in:
          type: integer
        total_tokens_out:
          type: integer
        total_cost_usd:
          type: number
        error_count:
          type: integer
          description: Count of failed spans in this trace
        metadata:
          type: object
        engine:
          $ref: '#/components/schemas/EngineTraceInfo'
    TraceDetail:
      allOf:
        - $ref: '#/components/schemas/Trace'
        - type: object
          properties:
            trace_id:
              type: string
            user_id:
              type: string
            tags:
              type: array
              items:
                type: string
            environment:
              type: string
            release:
              type: string
            input:
              description: Trace input payload (any valid JSON)
            output:
              description: Trace output payload (any valid JSON)
            engine:
              $ref: '#/components/schemas/EngineRunSummary'
...
    TraceList:
      type: object
      required: [traces, total]
      properties:
        traces:
          type: array
          items:
            $ref: '#/components/schemas/Trace'
        total:
          type: integer
    Span:
      type: object
      required: [id, trace_id, span_id, name, kind, status, started_at]
      properties:
        id:
          type: string
          format: uuid
        trace_id:
          type: string
          format: uuid
        span_id:
          type: string
          description: External span identifier (used for parent-child relationships)
        parent_span_id:
          type: string
          description: External parent span identifier
        name:
          type: string
        kind:
          type: string
          enum: [LLM, TOOL, CHAIN, AGENT, CUSTOM]
        status:
          type: string
          enum: [SCHEDULED, STARTED, COMPLETED, FAILED]
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        tokens_in:
          type: integer
        tokens_out:
          type: integer
        cost_usd:
          type: number
        latency_ms:
          type: integer
        error_message:
          type: string
        model:
          type: string
        provider:
          type: string
        input:
          description: Span input payload (any valid JSON)
        input_truncated:
          type: boolean
        input_original_size_bytes:
          type: integer
          format: int64
        input_truncation_reason:
          type: string
        output:
          description: Span output payload (any valid JSON)
        output_truncated:
          type: boolean
        output_original_size_bytes:
          type: integer
          format: int64
        output_truncation_reason:
          type: string
        metadata:
          type: object
    SpanList:
      type: object
      required: [spans]
      properties:
        spans:
          type: array
          items:
            $ref: '#/components/schemas/Span'
...
    TimelineEventType:
      type: string
      enum:
        [
          log,
          error,
          exception,
          message,
          metric,
          custom,
          state_change,
          decision,
          effect,
          wait,
          snapshot_marker,
          span_started,
          span_completed,
          span_failed,
        ]
    TimelineEventSource:
      type: string
      enum: [explicit, synthetic]
    TimelineEventLevel:
      type: string
      enum: [debug, info, warning, error]
    TimelineEvent:
      type: object
      required: [id, trace_id, event_type, timestamp, source]
      properties:
        id:
          type: string
          description: Opaque event identifier. Explicit events use persisted UUIDs; synthetic events use deterministic span-based IDs.
        trace_id:
          type: string
          format: uuid
        span_id:
          type: string
          description: External span identifier for the related span
        span_name:
          type: string
        event_type:
          $ref: '#/components/schemas/TimelineEventType'
        timestamp:
          type: string
          format: date-time
        source:
          $ref: '#/components/schemas/TimelineEventSource'
        level:
          $ref: '#/components/schemas/TimelineEventLevel'
        sequence:
          type: integer
          format: int32
          description: Explicit-event ordering value used to preserve same-timestamp ingest order. Absent for synthetic events.
        message:
          type: string
        payload:
          type: object
          description: Event payload details for explicit events and helper metadata for synthetic events
    TimelineResponse:
      type: object
      required: [events, trace_status, has_more]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/TimelineEvent'
        trace_status:
          type: string
          enum: [RUNNING, COMPLETED, FAILED]
        has_more:
          type: boolean
        next_cursor:
          type: string
          description: Opaque cursor for the next page when `has_more` is true
        poll_cursor:
          type: string
          description: Opaque cursor representing the last event included in this response. Use for incremental polling even when `has_more` is false.
        engine:
          type: object
          required: [projection_state]
          properties:
            projection_state:
              $ref: '#/components/schemas/EngineProjectionState'
...